% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns_store.R
\name{add_pattern}
\alias{add_pattern}
\title{Add a personal pattern to the patterns library}
\usage{
add_pattern(name, pattern_matrix, permanent = FALSE, user_path = NULL)
}
\arguments{
\item{name}{A character string giving the name of the pattern to be added.}

\item{pattern_matrix}{A binary matrix (containing only 0 and 1) representing the pattern.}

\item{permanent}{Logical. If \code{TRUE}, the patterns is added to the library and
permanently available for users. Defaults to \code{FALSE}: the pattern is available for the current session.}

\item{user_path}{internal}
}
\value{
A character vector of patterns available in the library during the current session.
}
\description{
The function allows to add temporarily or permanently a new pattern to the patterns library.
The names of all already available built-in patterns can be obtained with:
\code{list_patterns()}
}
\examples{
# temporary add a pattern built by the user
pattern <- matrix(
c(1,0,0,
1,1,1,
0,0,1),
nrow = 3,
byrow = TRUE)

add_pattern("my_own_pattern", pattern)

# permanently add a pattern extracted from a .rle file
\dontrun{
cordership <- read_rle("patterns/cordership.rle")

add_pattern("cordership", cordership, permanent = TRUE)
}
}
