% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns_store.R
\name{list_patterns}
\alias{list_patterns}
\title{List available Game of Life patterns}
\usage{
list_patterns()
}
\value{
A character vector of pattern names.
}
\description{
Returns the names of all built-in and permanently saved patterns available in
the package, as well as temporary patterns from the current session.
}
\examples{
list_patterns()
}
