% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{plot_grid}
\alias{plot_grid}
\title{Plotting binary (0, 1) matrices}
\usage{
plot_grid(grid)
}
\arguments{
\item{grid}{A binary (containing only 0 and 1) matrix or a character string
naming a built-in pattern or a list combining such matrices and/or pattern names.
Each element is plotted in a separate panel.}
}
\value{
Invisibly returns \code{NULL}. The function is called for its side effects
(plotting the matrix).
}
\description{
The function plots binary (containing only 0 and 1) matrix
}
\details{
Uses the internal library of patterns containing a named list of binary matrices representing a few common Game of Life patterns.
Available built-in pattern names can be obtained with \code{list_patterns()}.
The patterns library can be temporarily or permanently enriched with \code{add_pattern()}.
}
\examples{
# direct matrix
glider <- matrix(c(0,1,0,
0,0,1,
1,1,1),
nrow = 3,
byrow = TRUE)

plot_grid(glider)

# built-in matrix
plot_grid("pulsar")

# multiple patterns
glider <- matrix(c(0,1,0,
0,0,1,
1,1,1),
nrow = 3,
byrow = TRUE)

plot_grid(list("pulsar", glider))
}
