% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rle.R
\name{read_rle}
\alias{read_rle}
\title{Parsing .rle files into binary (0, 1) matrices}
\usage{
read_rle(rle_path)
}
\arguments{
\item{rle_path}{A character string for the relative or absolute path to an .rle file.}
}
\value{
A binary matrix (containing only 0 and 1)
}
\description{
The function supports standard Run Length Encoded (RLE) files commonly used
to describe Conway's Game of Life patterns.
The function reads an \code{.rle} file, extracts the RLE code and converts it
into a binary matrix (containing only 0 and 1).
}
\examples{
\dontrun{
glider <- read_rle("patterns/glider.rle")
image(glider)
}
}
